requirements.txt
-ip==21.1.2
absl-py==1.0.0
alabaster==0.7.12
antlr4-python3-runtime==4.8
astunparse==1.6.3
attrs==21.4.0
Babel==2.9.1
boto3==1.21.27
botocore==1.24.27
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
CMRESHandler==1.0.0
colorama==0.4.4
coverage==5.3.1
cryptography==36.0.2
cycler==0.11.0
deci-lab-client==2.15.1
Deprecated==1.2.13
docutils==0.17.1
elasticsearch==7.15.2
flatbuffers==2.0
fonttools==4.31.2
gast==0.5.3
google-auth-oauthlib==0.4.6
google-auth==2.6.0
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
hydra-core==1.1.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
Jinja2==3.1.1
jmespath==1.0.0
jsonschema==4.4.0
Keras-Preprocessing==1.1.2
keras==2.8.0
kiwisolver==1.4.0
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.1
onnx==1.11.0
onnxruntime==1.11.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pep517==0.12.0
Pillow==9.0.1
pip-tools==6.5.1
pip==21.2.4
protobuf==3.19.4
psutil==5.9.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyDeprecate==0.3.2
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==2.4.5
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests-oauthlib==1.3.1
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
scipy==1.8.0
setuptools==57.0.0
six==1.16.0
snowballstemmer==2.2.0
sphinx-rtd-theme==1.0.0
Sphinx==4.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
super-gradients==1.7.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow==2.8.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tomli==2.0.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchmetrics==0.7.3
torchvision==0.12.0+cu113
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wheel==0.36.2
wrapt==1.14.0
zipp==3.7.0
